Clarify how to use ``--config`` option
I was struggling for the longest time to get the ``--config`` option working. After much hair-pulling, I worked out that it needs to be passed as an option to the ``fix`` command, and not to the ``php-cs-fixer`` command directly.

This change will hopefully make that clear to other users and help them avoid the same headache.